+++ src/DataObjects/CreateCustomerData.php
@@ -9,7 +9,7 @@ use Subster\PhpSdk\Concerns\Data;
 class CreateCustomerData extends Data
 {
     public function __construct(
-        public readonly string $email,
+        public readonly ?string $email = null,
         public readonly ?string $name = null,
     ) {}
 }